requirements.txt
# Core dependencies
python-telegram-bot==20.7
requests==2.31.0
python-dotenv==1.0.0

# Optional - if you need additional functionality
asyncio-mqtt==0.16.1